Move price component
