+++ src/components/trade/Price.tsx
